backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic
python-multipart
google-generativeai